+++ main.py
@@ -449,7 +449,7 @@ class MainWindow(QMainWindow):
 
         cp = np.array([
             [[0, 0, 0], [10, 10, 0], [5, 0, 0], [15, 10, 0]]
-        ], dtype=float).transpose()
+        ]).transpose()
 
         geometry = HermiteCurve(cp, number_u)
         self.add_geometry(geometry)
@@ -491,7 +491,7 @@ class MainWindow(QMainWindow):
             [[10,0,0], [10,10,0], [10,1,0], [10,11,0]],
             [[1,0,0], [1,10,0], [0,0,1], [0,10,1]],
             [[11,0,0], [11,10,0], [10,0,1], [10,10,1]],
-        ], dtype=float).transpose((2, 0, 1))
+        ]).transpose((2, 0, 1))
 
         geometry = HermiteSurface(cp, number_u, number_v)
         self.add_geometry(geometry)
